backend/requirements.txt
# ==========================================
# 核心框架
# ==========================================
Django==4.2.8
djangorestframework==3.14.0

# ==========================================
# 数据库
# ==========================================
mysqlclient==2.2.0

# ==========================================
# 认证和安全
# ==========================================
PyJWT==2.8.0
django-cors-headers==4.3.1

# ==========================================
# 配置管理
# ==========================================
python-dotenv==1.0.0

# ==========================================
# API 功能
# ==========================================
django-filter==23.5
drf-spectacular==0.27.0

# ==========================================
# 文件处理
# ==========================================
Pillow==10.1.0
django-ckeditor==6.7.0

# ==========================================
# 性能优化
# ==========================================
django-redis==5.4.0
redis==5.0.1

# ==========================================
# 监控和错误追踪（可选）
# ==========================================
# sentry-sdk==1.39.2

# ==========================================
# 测试
# ==========================================
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# ==========================================
# 代码质量
# ==========================================
flake8==6.1.0
black==23.12.1
mypy==1.7.1
django-stubs==4.2.7

# ==========================================
# 开发工具
# ==========================================
ipython==8.12.0
django-debug-toolbar==4.2.0

# ==========================================
# 生产环境
# ==========================================
gunicorn==21.2.0
whitenoise==6.6.0

